feat(auth): add email verification resend functionality
Implements a new endpoint to allow users to request a resend of the 
email verification link. This feature enhances user experience by 
providing a way for users who did not receive the initial email to 
verify their accounts. The endpoint is designed to prevent email 
enumeration by always returning a success message, regardless of 
whether the email exists or is verified. Additionally, it includes 
rate limiting for security.
