+++ routes/api_v0.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 use App\Http\Controllers\Api\V0\AuthController;
 use App\Http\Controllers\Api\V0\PingController;
-use App\Http\Controllers\Api\V0\VerificationController;
 use Illuminate\Support\Facades\Route;
 
 // Simple ping endpoint to check API health
@@ -13,12 +12,7 @@ Route::get('/ping', PingController::class)->name('api.v0.ping');
 // Authentication
 Route::post('/auth/login', [AuthController::class, 'login'])->name('api.v0.auth.login');
 Route::post('/auth/register', [AuthController::class, 'register'])->name('api.v0.auth.register');
-
-// Email Verification Handling
-Route::get('/auth/email/verify/{id}/{hash}', [VerificationController::class, 'verify'])
-    ->middleware(['signed', 'throttle:6,1'])
-    ->name('api.v0.auth.verify');
-Route::post('/auth/email/resend', [VerificationController::class, 'resend'])
+Route::post('/auth/email/resend', [AuthController::class, 'resend'])
     ->middleware('throttle:3,60')
     ->name('api.v0.auth.resend');
