+++ tests/Feature/Api/V0/Auth/ResendVerificationTest.php
@@ -4,94 +4,11 @@ declare(strict_types=1);
 
 namespace Tests\Feature\Api\V0;
 
-use App\Enums\Api\V0\ApiErrorCode;
 use App\Models\User;
 use App\Notifications\VerifyEmail;
 use Illuminate\Auth\Events\Verified;
-use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Notification;
-use Illuminate\Support\Facades\URL;
 use Symfony\Component\HttpFoundation\Response;
-
-it('verifies a user with a valid signed url', function () {
-    Event::fake();
-
-    $user = User::factory()->create(['email_verified_at' => null]);
-    $verificationUrl = URL::temporarySignedRoute(
-        'api.v0.auth.verify',
-        now()->addMinutes(60),
-        ['id' => $user->id, 'hash' => sha1($user->getEmailForVerification())]
-    );
-    $response = $this->getJson($verificationUrl);
-
-    $response
-        ->assertStatus(Response::HTTP_OK)
-        ->assertExactJson([
-            'success' => true,
-            'data' => [
-                'message' => 'Email verified successfully.',
-            ],
-        ]);
-
-    $user->refresh();
-
-    expect($user->hasVerifiedEmail())->toBeTrue();
-    Event::assertDispatched(Verified::class, fn ($e) => $e->user->id === $user->id);
-});
-
-it('returns error if verification url signature is invalid', function () {
-    $user = User::factory()->create(['email_verified_at' => null]);
-    $baseUrl = URL::route(
-        'api.v0.auth.verify',
-        ['id' => $user->id, 'hash' => sha1($user->getEmailForVerification())],
-        absolute: false
-    );
-    $urlWithBadSignature = $baseUrl.'?expires=123&signature=badsignature';
-
-    $response = $this->getJson($urlWithBadSignature);
-
-    $response->assertStatus(Response::HTTP_FORBIDDEN);
-    expect($user->refresh()->hasVerifiedEmail())->toBeFalse();
-});
-
-it('returns error if verification url hash is invalid', function () {
-    $user = User::factory()->create(['email_verified_at' => null]);
-    $verificationUrl = URL::temporarySignedRoute(
-        'api.v0.auth.verify',
-        now()->addMinutes(60),
-        ['id' => $user->id, 'hash' => 'invalid-hash']
-    );
-
-    $response = $this->getJson($verificationUrl);
-
-    $response
-        ->assertStatus(Response::HTTP_BAD_REQUEST)
-        ->assertExactJson([
-            'success' => false,
-            'code' => ApiErrorCode::VERIFICATION_INVALID->value,
-            'message' => 'Invalid or expired verification link.',
-        ]);
-    expect($user->refresh()->hasVerifiedEmail())->toBeFalse();
-});
-
-it('returns error if user is already verified', function () {
-    $user = User::factory()->create(['email_verified_at' => now()]);
-    $verificationUrl = URL::temporarySignedRoute(
-        'api.v0.auth.verify',
-        now()->addMinutes(60),
-        ['id' => $user->id, 'hash' => sha1($user->getEmailForVerification())]
-    );
-
-    $response = $this->getJson($verificationUrl);
-
-    $response
-        ->assertStatus(Response::HTTP_BAD_REQUEST)
-        ->assertExactJson([
-            'success' => false,
-            'code' => ApiErrorCode::ALREADY_VERIFIED->value,
-            'message' => 'Email already verified.',
-        ]);
-});
 
 it('allows resend request for unverified user and sends notification', function () {
     Notification::fake();
